+++ e2e/testcases/form-state/form-state.spec.ts
@@ -6,20 +6,17 @@ import {
   getToken,
   goToSection,
   selectDeclarationAction
-} from '../../../helpers'
+} from '../../helpers'
 import { faker } from '@faker-js/faker'
-import { fillChildDetails, openBirthDeclaration } from '../../birth/helpers'
-import { CLIENT_URL, CREDENTIALS } from '../../../constants'
-import {
-  createDeclaration,
-  Declaration
-} from '../../test-data/birth-declaration'
-import { selectAction, type } from '../../../utils'
+import { fillChildDetails, openBirthDeclaration } from '../birth/helpers'
+import { CLIENT_URL, CREDENTIALS } from '../../constants'
+import { createDeclaration, Declaration } from '../test-data/birth-declaration'
+import { selectAction, type } from '../../utils'
 import {
   navigateToCertificatePrintAction,
   selectRequesterType
-} from '../../print-certificate/birth/helpers'
-import { REQUIRED_VALIDATION_ERROR } from '../../birth/helpers'
+} from '../print-certificate/birth/helpers'
+import { REQUIRED_VALIDATION_ERROR } from '../birth/helpers'
 
 test.describe('Form state', () => {
   test.describe
