+++ e2e/testcases/form-state/informant-reset.spec.ts
@@ -1,7 +1,7 @@
 import { Page, expect, test } from '@playwright/test'
-import { goToSection, login } from '../../../helpers'
-import { openBirthDeclaration } from '../../birth/helpers'
-import { CREDENTIALS } from '../../../constants'
+import { goToSection, login } from '../../helpers'
+import { openBirthDeclaration } from '../birth/helpers'
+import { CREDENTIALS } from '../../constants'
 
 test.describe('Informant details resets when relation is changed', () => {
   test.describe.serial('Birth', async () => {
